Rename insertion.rs to insertion_sort.rs
Renamed the file and affected references to follow the current naming
conventions and to fix the broken insertion sort file link in the README.
