+++ src/sorting/mod.rs
@@ -1,7 +1,7 @@
 mod bubble_sort;
 mod counting_sort;
 mod heap_sort;
-mod insertion;
+mod insertion_sort;
 mod merge_sort;
 mod quick_sort;
 mod radix_sort;
@@ -13,7 +13,7 @@ pub use self::bubble_sort::bubble_sort;
 pub use self::counting_sort::counting_sort;
 pub use self::counting_sort::generic_counting_sort;
 pub use self::heap_sort::heap_sort;
-pub use self::insertion::insertion_sort;
+pub use self::insertion_sort::insertion_sort;
 pub use self::merge_sort::merge_sort;
 pub use self::quick_sort::quick_sort;
 pub use self::radix_sort::radix_sort;
